Component Logic Modularized to simplify code
